requirements.txt
asgiref==3.2.10
Django==3.0.8
django-environ==0.4.5
gunicorn==20.0.4
pytz==2020.1
sqlparse==0.3.1

djangorestframework~=3.13.1
django-filter~=21.1